% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_wide_table.R
\name{as_wide_table}
\alias{as_wide_table}
\title{Create a single wide table from the AusTraits data object}
\usage{
as_wide_table(austraits)
}
\arguments{
\item{austraits}{austraits data object}
}
\value{
A single wide table with collapsed contexts and sites text and with
some cols renamed for alignment with other reosurces
}
\description{
Create a single wide table from the AusTraits data object
}
\examples{
\dontrun{
data <- austraits
data \%>\% as_wide_table()
}
}
