% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austraits-package.R
\docType{package}
\name{austraits-package}
\alias{austraits}
\alias{austraits-package}
\title{austraits: Helpful functions to access the AusTraits database and wrangle data from other traits.build databases}
\description{
`austraits` allow users to **access, explore and wrangle data** from traits.build relational databases. It is also an R interface to AusTraits, the Australian plant trait database. This package contains functions for joining data from various tables, filtering to specific records, combining multiple databases and visualising the distribution of the data. We expect this package will assist users in working with `traits.build` databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://traitecoevo.github.io/austraits/}
  \item Report bugs at \url{https://github.com/traitecoevo/austraits/issues}
}

}
\author{
\strong{Maintainer}: Fonti Kar \email{f.kar@unsw.edu.au} (\href{https://orcid.org/0000-0002-2760-3974}{ORCID})

Authors:
\itemize{
  \item Daniel Falster \email{daniel.falster@unsw.edu.au} (\href{https://orcid.org/0000-0002-9814-092X}{ORCID})
  \item Dony Indiarto (\href{https://orcid.org/0000-0001-9546-8201}{ORCID})
}

Other contributors:
\itemize{
  \item Elizabeth Wenk \email{e.wenk@unsw.edu.au} (\href{https://orcid.org/0000-0001-5640-5910}{ORCID}) [contributor]
}

}
\keyword{internal}
