% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australia_map_raster.R
\docType{data}
\name{australia_map_raster}
\alias{australia_map_raster}
\title{A raster map of Australia}
\format{
A RasterLayer 692 x 814, 563288 cells
}
\source{
https://dbr.abs.gov.au/absmaps/index.html
}
\usage{
australia_map_raster
}
\description{
A raster object to spatially plot traits onto map of Australia
}
\keyword{datasets}
