% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australia_raster.R
\docType{data}
\name{australia_raster}
\alias{australia_raster}
\title{A dataframe containing raster coordinates of Australia}
\format{
250000 obs
}
\source{
see data-raw create_raster.R
}
\usage{
australia_raster
}
\description{
A dataframe containing raster coordinates of Australia, used in plot_site_locations
}
\keyword{datasets}
