% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_databases.R
\name{bind_databases}
\alias{bind_databases}
\title{Bind multiple traits.build data objects into a single data object}
\usage{
bind_databases(database_1, ...)
}
\arguments{
\item{database_1}{List of traits.build databases to bind together}

\item{...}{Arguments passed to other functions}
}
\value{
Compiled database as a single large list
}
\description{
\code{bind_databases} binds all the listed studies into a single traits.build
database object as a large list.
}
