% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_trait_values.R
\name{bind_trait_values}
\alias{bind_trait_values}
\title{Bind trait values}
\usage{
bind_trait_values(trait_data)
}
\arguments{
\item{trait_data}{The trait data frame generated from austraits - see example}
}
\value{
tibble that is condensed down where multiple observations in value, value_type and replicates are collapsed down and separated by '--'
}
\description{
This function condenses data for studies that have multiple observations for a given trait into a single row.
This function concatenates multiple values into a single cell
}
\examples{
\dontrun{
traits <- austraits$traits \%>\% 
dplyr::filter(dataset_id == "ABRS_1981")
traits
traits_bind <- bind_trait_values(traits)
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
