% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{check_compatibility}
\alias{check_compatibility}
\title{Check compatibility of traits.build object}
\usage{
check_compatibility(database, single_table_allowed = FALSE)
}
\arguments{
\item{database}{traits.build database (list object)}

\item{single_table_allowed}{logical for when the input might be a single table instead of a complete database; defaults to FALSE}
}
\value{
logical (T/F) output and messaging for uncompatible versions
}
\description{
Function to check whether the data object has been compiled by the traits.build workflow and
therefore has a data structure that is appropriate for use with austraits functions.
}
\examples{
\dontrun{
check_compatibility(database)
}
}
\author{
Elizabeth Wenk - e.wenk@unsw.edu.au
}
\keyword{internal}
