% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{check_traits_compatibility}
\alias{check_traits_compatibility}
\title{Check compatibility of traits table}
\usage{
check_traits_compatibility(traits)
}
\arguments{
\item{traits}{table in traits.build object}
}
\value{
logical, TRUE indicating version traits table came from traits.build version > 1.0
}
\description{
Check compatibility of traits table
}
\keyword{internal}
