% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_list_to_df1}
\alias{convert_list_to_df1}
\title{Convert list with single entries to dataframe}
\usage{
convert_list_to_df1(my_list)
}
\arguments{
\item{my_list}{A list with single entries}
}
\value{
A tibble with two columns
}
\description{
Convert a list with a single level of entries to a dataframe,
useful when converting a yaml into a dataframe.
}
\examples{
\dontrun{
convert_list_to_df1(as.list(dplyr::starwars)[2])
}
}
