% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{create_metadata}
\alias{create_metadata}
\title{Helper function to create nice metadata table}
\usage{
create_metadata(res)
}
\arguments{
\item{res}{output of austraits.json}
}
\value{
dataframe of metadata (date of release, doi and version)
}
\description{
Helper function to create nice metadata table
}
