% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_create_combined_table.R
\name{database_create_combined_table}
\alias{database_create_combined_table}
\title{Create combined traits.build table}
\usage{
database_create_combined_table(database)
}
\arguments{
\item{database}{A traits.build database}
}
\value{
A table combining information in 7 traits.build relational tables: traits, locations, contexts, methods, taxa, taxonomic_updates, and contributors
}
\description{
Create a single database output that merges together the information
in all relational tables within a traits.build database.
Trait measurements are still output in long format (1 row per trait value),
but all measurement-related metadata (methods, location properties, context properties, contributors)
are now included as additional columns in a single table.
}
