% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{download_austraits}
\alias{download_austraits}
\title{Function for loading .rds AusTraits files}
\usage{
download_austraits(url, filename, path)
}
\arguments{
\item{url}{url of download via Zenodo API}

\item{filename}{Name of file that will be downloaded e.g. austraits-3.0.2.rds}

\item{path}{file path to where AusTraits will be downloaded}
}
\description{
Function for loading .rds AusTraits files
}
