% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract data from traits.build database}
\usage{
extract_data(
  database,
  table = NA,
  col,
  col_value,
  partial_matches_allowed = TRUE
)
}
\arguments{
\item{database}{traits.build database (list object)}

\item{table}{Table within a traits.build database}

\item{col}{Column name within the specified table.}

\item{col_value}{Value (of column, from with a table) that is used to subset database. This can be a single value or a vector. It includes partial string matches.}
}
\value{
subset traits.build database
}
\description{
Function to extract data from a traits.build database based on
any value(s) from any column in the traits, locations, contexts, methods,
taxa, taxonomic_updates, and identifiers tables.
The output a traits.build formatted database with all tables subset
based on the specified table, column (variable) and column value.
}
\examples{
\dontrun{
extract_data(database = traits.build_database, table = "traits", 
col = "trait_name", col_value = "leaf_area")
}
}
