% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxa.R
\name{extract_taxa}
\alias{extract_taxa}
\title{Extract data for one specific taxa}
\usage{
extract_taxa(austraits, family = NULL, genus = NULL, taxon_name = NULL)
}
\arguments{
\item{austraits}{austraits list object}

\item{family}{character string of family}

\item{genus}{character string of genus}

\item{taxon_name}{character string of taxon name}
}
\value{
A large list of tibbles containing all austraits information for specificied taxa
}
\description{
Function to subset of all data associated with a particular dataset from austraits
}
\examples{
\dontrun{
extract_taxa(austraits, family = "Proteaceae")
extract_taxa(austraits, genus = "Acacia")
}
}
\author{
Fonti Kar - f.kar@unsw.edu.au
}
