% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trait.R
\name{extract_trait}
\alias{extract_trait}
\title{Extract all data for specific traits}
\usage{
extract_trait(database, trait_names, taxon_names)
}
\arguments{
\item{database}{a large list of tibbles built by \code{traits.build}}

\item{trait_names}{character string of trait(s) for which data will be extracted}

\item{taxon_names}{optional argument, specifying taxa for which data will be extracted}
}
\value{
List of tibbles containing all traits.build data and metadata for the specified trait(s).
}
\description{
Function to subset all data associated with a particular trait from a traits.build relational database.
}
\details{
\code{extract_trait} has been developed to extract data for specific traits from databases built using the traits.build workflow.
Learn more at:
\url{https://github.com/traitecoevo/traits.build} &
\url{https://github.com/traitecoevo/traits.build-book}

Note to AusTraits users:
\itemize{
\item This function works with AusTraits version >= 5.0.0 (from Nov 2023 release)
\item For AusTraits versions <= 4.2.0 (up to Sept 2023 release) see \url{https://github.com/traitecoevo/austraits}  for how to install old versions of the package or download a newer version of the database.
}
}
\examples{
\dontrun{
extract_trait(database, "wood_density", taxon_name = "Acacia celsa")
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
