% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_database.R
\name{flatten_database}
\alias{flatten_database}
\title{Create combined traits.build table}
\usage{
flatten_database(database)
}
\arguments{
\item{database}{A traits.build database}

\item{format}{}

\item{vars}{list}

\item{include_description}{logical}
}
\value{
A table combining information in 7 traits.build relational tables: traits, locations, contexts, methods, taxa, taxonomic_updates, and contributors
}
\description{
Create a single database output that merges together the information
in all relational tables within a traits.build database.
Trait measurements are still output in long format (1 row per trait value),
but all measurement-related metadata (methods, location properties, context properties, contributors)
are now included as additional columns in a single table.
}
