% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{get_compiled_by_traits.build}
\alias{get_compiled_by_traits.build}
\title{Retrieve compiled by information from metadata table}
\usage{
get_compiled_by_traits.build(database)
}
\arguments{
\item{database}{traits.build database}
}
\value{
logical, TRUE indicating version traits table came from traits.build version > 1.0
}
\description{
Retrieve compiled by information from metadata table
}
