% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_traits_table}
\alias{get_traits_table}
\title{Retrieve traits table if user passes traits.build object.}
\usage{
get_traits_table(aus_traits)
}
\arguments{
\item{aus_traits}{traits.build object or traits table}
}
\description{
Retrieve traits table if user passes traits.build object.
}
