% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{get_version_latest}
\alias{get_version_latest}
\title{Retrieve the latest version of AusTraits}
\usage{
get_version_latest(path = "data/austraits", update = TRUE)
}
\arguments{
\item{path}{file path to where AusTraits will be downloaded}

\item{update}{if TRUE, AusTraits versions .json will be re-downloaded}
}
\value{
character string of latest version
}
\description{
Retrieve the latest version of AusTraits
}
