% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{get_versions}
\alias{get_versions}
\title{Print out AusTraits versions}
\usage{
get_versions(path, update = TRUE)
}
\arguments{
\item{path}{A file path where AusTraits was previously downloaded}

\item{update}{Would you like the versions json be updated in case of new releases?}
}
\value{
A tibble containing version numbers and doi which can be used in load_austraits()
}
\description{
Print out AusTraits versions
}
\examples{
\dontrun{
austraits <- load_austraits(version = "3.0.2", path = "data/austraits")
}
}
