% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_location_coordinates}
\alias{join_location_coordinates}
\title{Joining location coordinates to traits table}
\usage{
join_location_coordinates(austraits)
}
\arguments{
\item{austraits}{traits.build generated database}

\item{vars}{Vector specifying which columns or values from a specific relational table to join to the traits table (currently works only for \code{join_} functions joining a single dataframe). For the functions \code{join_methods}, \code{join_taxa}, \code{join_taxonomic_updates}, and \code{join_contributors} the parameter vars specifies which columns to join to the traits table. For the functions \code{join_location_properties} and \code{join_context_properties} the parameter \code{vars} specifies which properties to add to the traits table; the functions \code{lookup_location_property} and \code{lookup_context_property} allow easy searches for relevant properties. The function \code{join_location_coordinates} does not include the parameter \code{vars}.}

\item{format}{A parameter for \code{join_context_properties}, \code{join_location_properties}, and \code{join_contributors}. Specifies whether location properties or context properties for a given observation will be concatenated, and if concatenated whether the compacted column(s) will have a human readable ("single_column_pretty") format or be format using json ("single_column_json") syntax. For location and context properties there is also the option to add each location_property or \code{context_property} to the traits table as its own column ("many_columns").}

\item{include_description}{For \code{join_context_properties} only, a logical indicating whether to include (TRUE) or omit (FALSE) the context_property descriptions.}
}
\value{
traits.build list object, but with additional fields (columns) appended to \code{traits} dataframe
}
\description{
Function to merge metadata stored in relational tables into the core traits table.

A traits.build database includes relational tables that contain measurement (and observation) metadata pertaining to locations, contexts, methods, and taxonomy. This function merges all (or some) of the metadata from one (or all) of these tables into the traits table.
}
\details{
the \code{join_} functions have been developed to join relational tables for databases built using the traits.build workflow.
Learn more at:
\url{https://github.com/traitecoevo/traits.build} &
\url{https://github.com/traitecoevo/traits.build-book}

Note to AusTraits users:
\itemize{
\item This function works with AusTraits version >= 5.0.0 (from Nov 2023 release)
\item For AusTraits versions <= 4.2.0 (up to Sept 2023 release) see \url{https://github.com/traitecoevo/austraits}  for how to install old versions of the package or download a newer version of the database.
}
}
\examples{
\dontrun{
austraits$traits

#Append locations coordinates and output a single table
(austraits \%>\% join_location_coordinates)$traits

#Append location coordinates, but maintain the relational database
austraits \%>\% join_location_coordinates()

#Append contexts
(austraits \%>\% join_context_properties)$traits

#Append methods
(austraits \%>\% join_methods)$traits

#Append taxonomic details
(austraits \%>\% join_taxa)$traits

#Append taxonomic update details
(austraits \%>\% join_taxonomic_updates)$traits
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
