% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_all}
\alias{join_all}
\alias{join_taxonomy}
\alias{join_taxonomic_updates}
\alias{join_methods}
\alias{join_contributors}
\alias{join_location_coordinates}
\alias{join_location_properties}
\title{Join metadata fields into the core traits table}
\usage{
join_all(austraits)

join_taxonomy(
  austraits,
  vars = c("family", "genus", "taxon_rank", "establishment_means")
)

join_taxonomic_updates(austraits, vars = c("original_name"))

join_methods(austraits, vars = c("methods"))

join_contributors(austraits, format = "single_column_pretty", vars = "all")

join_location_coordinates(austraits)

join_location_properties(
  austraits,
  format = "single_column_pretty",
  vars = "all"
)
}
\arguments{
\item{austraits}{traits.build generated database}

\item{vars}{vector specifying which columns from a specific relational table to join to the traits table (works only for \code{join_} functions joining a single dataframe)}

\item{format}{for \code{join_context_properties} and \code{join_location_properties} only. Logical, whether location properties or context properties for a given observation will be concatenated into one cell}
}
\value{
traits.build list object, but with additional fields (columns) appended to \code{traits} dataframe
}
\description{
Function to merge metadata stored in relational tables into the core traits table.

A traits.build database includes relational tables that contain measurement (and observation) metadata pertaining to locations, contexts, methods, and taxonomy. This function merges all (or some) of the metadata from one (or all) of these tables into the traits table.
}
\details{
the \code{join_} functions have been developed to join relational tables for databases built using the traits.build workflow.
Learn more at:
\link{https://github.com/traitecoevo/traits.build} &
\link{https://github.com/traitecoevo/traits.build-book}

Note to AusTraits users:
\itemize{
\item This function works with AusTraits version >= 5.0.0 (from Nov 2023 release)
\item For AusTraits versions <= 4.2.0 (up to Sept 2023 release) see \link{https://github.com/traitecoevo/austraits} for how to install old versions of the package or download a newer version of the database."
}
}
\examples{
\dontrun{
austraits$traits

#Append locations coordinates and output a single table
(austraits \%>\% join_location_coordinates)$traits

#Append location coordinates, but maintain the relational database
austraits \%>\% join_location_coordinates()

#Append contexts
(austraits \%>\% join_context_properties)$traits

# Append methods
(austraits \%>\% join_methods(vars = c("method_id")))$traits

#Append taxonomic details
(austraits \%>\% join_taxonomy)$traits

#Append all information
(austraits \%>\% join_all)$traits
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
