% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_context_properties}
\alias{join_context_properties}
\title{Joining context properties to traits table}
\usage{
join_context_properties(
  austraits,
  format = "single_column_pretty",
  vars = "all",
  include_description = TRUE
)
}
\description{
Joining context properties to traits table
}
