% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_contexts1}
\alias{join_contexts1}
\title{Joining location info for AusTraits versions <= 3.0.2}
\usage{
join_contexts1(
  austraits,
  vars = c("dataset_id", "context_name", "context_property", "value")
)
}
\description{
Joining location info for AusTraits versions <= 3.0.2
}
