% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_contexts2}
\alias{join_contexts2}
\title{Joining location info for AusTraits versions > 3.0.2}
\usage{
join_contexts2(austraits, ...)
}
\description{
Joining location info for AusTraits versions > 3.0.2
}
