% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_.R
\name{join_identifiers}
\alias{join_identifiers}
\title{Joining identifiers to traits table}
\usage{
join_identifiers(database, format = "single_column_pretty", vars = "all")
}
\arguments{
\item{database}{traits.build database (list object)}

\item{format}{Specifies whether metadata from the identifiers is output in a human readable format ("single_column_pretty"; default), with each location property added as a separate column ("many_columns") or using json syntax ("single_column_json").}

\item{vars}{Identifier types for which data is to be appended to the traits table, defaulting to all identifier types (vars = "all").}
}
\value{
traits.build list object, but with identifiers from the identifiers table appended to the traits table.
}
\description{
Function to merge metadata from the identifiers table of a traits.build database into the core traits table.
}
\details{
the \code{join_} functions have been developed to join relational tables for databases built using the traits.build workflow.
Learn more at:
\url{https://github.com/traitecoevo/traits.build} &
\url{https://github.com/traitecoevo/traits.build-book}

Note to AusTraits users:
\itemize{
\item This function works with AusTraits version >=7.0.0 (2025 releases)
\item For AusTraits versions <= 4.2.0 (up to Sept 2023 release) see \url{https://github.com/traitecoevo/austraits}  for how to install old versions of the package or download a newer version of the database.
}
}
\examples{
\dontrun{
(database \%>\% join_identifiers(format = "single_column_pretty", vars = "all"))$traits
}
}
