% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_taxa}
\alias{join_taxa}
\alias{join_taxonomic_updates}
\alias{join_methods}
\alias{join_contributors}
\alias{join_location_properties}
\title{Joining taxonomy to traits table}
\usage{
join_taxa(
  austraits,
  vars = c("family", "genus", "taxon_rank", "establishment_means")
)

join_taxonomic_updates(austraits, vars = c("aligned_name"))

join_methods(austraits, vars = c("methods"))

join_contributors(austraits, format = "single_column_pretty", vars = "all")

join_location_properties(
  austraits,
  format = "single_column_pretty",
  vars = "all"
)
}
\description{
Joining taxonomy to traits table

Joining taxonomic updates information to traits table

Joining methodological information to traits table

Joining data contributor metadata to traits table

Joining location properties to traits table
}
