% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_locations2}
\alias{join_locations2}
\title{Joining location info for AusTraits versions > 3.0.2}
\usage{
join_locations2(austraits, vars = c("longitude (deg)", "latitude (deg)"))
}
\description{
Joining location info for AusTraits versions > 3.0.2
}
