% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_.R
\name{join_taxa}
\alias{join_taxa}
\title{Joining taxonomy to traits table}
\usage{
join_taxa(
  database,
  vars = c("family", "genus", "taxon_rank", "establishment_means")
)
}
\arguments{
\item{database}{traits.build database (list object)}

\item{vars}{Columns from the taxa table to be joined to the traits table, defaulting to c("family", "genus", "taxon_rank", "establishment_means").}
}
\value{
traits.build list object, but with additional fields (columns) for the specified variables from the taxa table appended to the traits table.
}
\description{
Function to merge metadata from the taxa table of a traits.build database into the core traits table.
}
\details{
the \code{join_} functions have been developed to join relational tables for databases built using the traits.build workflow.
Learn more at:
\url{https://github.com/traitecoevo/traits.build} &
\url{https://github.com/traitecoevo/traits.build-book}

Note to AusTraits users:
\itemize{
\item This function works with AusTraits version >= 5.0.0 (from Nov 2023 release)
\item For AusTraits versions <= 4.2.0 (up to Sept 2023 release) see \url{https://github.com/traitecoevo/austraits}  for how to install old versions of the package or download a newer version of the database.
}
}
\examples{
\dontrun{
#Append taxonomic details
(database \%>\% join_taxa)$traits
}
}
