% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_.R
\name{join_taxonomic_updates}
\alias{join_taxonomic_updates}
\title{Joining taxonomic updates information to traits table}
\usage{
join_taxonomic_updates(database, vars = c("aligned_name"))
}
\arguments{
\item{database}{traits.build database (list object)}

\item{vars}{Columns from the taxa table to be joined to the traits table, defaulting to c("aligned_name").}
}
\value{
traits.build list object, but with additional fields (columns) for the specified variables from the taxonomic_updates table appended to the traits table.
}
\description{
Function to merge metadata from the taxonomic_updates table of a traits.build database into the core traits table.
}
\details{
the \code{join_} functions have been developed to join relational tables for databases built using the traits.build workflow.
Learn more at:
\url{https://github.com/traitecoevo/traits.build} &
\url{https://github.com/traitecoevo/traits.build-book}

Note to AusTraits users:
\itemize{
\item This function works with AusTraits version >= 5.0.0 (from Nov 2023 release)
\item For AusTraits versions <= 4.2.0 (up to Sept 2023 release) see \url{https://github.com/traitecoevo/austraits}  for how to install old versions of the package or download a newer version of the database.
}
}
\examples{
\dontrun{
#Append taxonomic update details
(database \%>\% join_taxonomic_updates)$traits
}
}
