% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_taxonomy1}
\alias{join_taxonomy1}
\title{Joining taxonomic info for AusTraits versions <= 3.0.2}
\usage{
join_taxonomy1(
  austraits,
  vars = c("family", "genus", "taxonRank", "acceptedNameUsageID")
)
}
\description{
Joining taxonomic info for AusTraits versions <= 3.0.2
}
