% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_taxonomy2}
\alias{join_taxonomy2}
\title{Joining taxonomic info for AusTraits versions > 3.0.2}
\usage{
join_taxonomy2(
  austraits,
  vars = c("family", "genus", "taxon_rank", "accepted_name_usage_id")
)
}
\description{
Joining taxonomic info for AusTraits versions > 3.0.2
}
