% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{list1_to_df}
\alias{list1_to_df}
\title{Convert a list with single entries to dataframe}
\usage{
list1_to_df(my_list)
}
\arguments{
\item{my_list}{a list with single entries}
}
\value{
a tibble with two columns
}
\description{
Convert a list with single entries to dataframe
}
\examples{
list1_to_df(as.list(iris)[2])
}
\keyword{internal}
