% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{load_austraits}
\alias{load_austraits}
\alias{load_austraits_lite}
\title{Load AusTraits}
\usage{
load_austraits(path, link, update)

load_austraits_lite()
}
\arguments{
\item{path}{file path to where download data. By default set to "data/austraits"}

\item{link}{link to Zenodo record to download}

\item{update}{should files be downloaded anew?}
}
\value{
AusTraits data object


}
\description{
Downloads AusTraits version 3.0.2 and reads into R

Load lite version of austraits for demo and testing purposes
}
\examples{
\dontrun{
austraits_lite <- load_austraits_lite() #loads Lite version for testing
austraits <- load_austraits()
}
}
