% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{load_json}
\alias{load_json}
\title{Load the austraits.json}
\usage{
load_json(path, update)
}
\arguments{
\item{path}{file path to where AusTraits will be downloaded}

\item{update}{if TRUE, AusTraits versions .json will be re-downloaded}
}
\description{
Load the austraits.json
}
