% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_.R
\name{lookup_context_property}
\alias{lookup_context_property}
\title{Look up context properties}
\usage{
lookup_context_property(database, term)
}
\arguments{
\item{database}{traits.build database (list object)}

\item{term}{character string for context property search term}
}
\value{
vector containing context properties that contains search term
}
\description{
Look up context properties that contain a specific search term.
}
\examples{
\dontrun{
austraits \%>\% lookup_context_property("temperature")
}
}
