% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{my_kable_styling_html}
\alias{my_kable_styling_html}
\title{html styling for kable tables}
\usage{
my_kable_styling_html(...)
}
\arguments{
\item{...}{arguments passed on to kableExtra::kable, kableExtra::kable_styling}
}
\description{
html styling for kable tables
}
\keyword{internal}
