% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{my_kable_styling_pdf}
\alias{my_kable_styling_pdf}
\title{pdf styling for kable tables}
\usage{
my_kable_styling_pdf(...)
}
\arguments{
\item{...arguments}{passed on to kableExtra::kable}
}
\description{
pdf styling for kable tables
}
\keyword{internal}
