% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locations.R
\name{plot_locations}
\alias{plot_locations}
\alias{plot_site_locations}
\title{Produce location maps of trait values}
\usage{
plot_locations(aus_traits, feature = "trait_name", ...)
}
\arguments{
\item{aus_traits}{austraits object OR traits table. Note location details must be joined. See join_all and examples}

\item{feature}{grouping/classification categories e.g trait_name, collection_type for <= v3.0.2, basis of record for >3.0.2}

\item{...}{arguments passed to ggplot()}
}
\value{
ggplot of sites
}
\description{
Plot location where trait data was collected from
}
\examples{
\dontrun{
#All traits from a given study
data <- austraits \%>\% extract_dataset(dataset_id = "Falster_2003") \%>\% join_all() 
data \%>\% plot_locations("trait_name")

#Single trait
data <- austraits \%>\% extract_trait(trait_names = c("plant_height")) \%>\% join_all() 
data$traits \%>\% plot_locations("trait_name")
}
}
\author{
Dony Indiarto - d.indiarto@student.unsw.edu.au
}
