% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.austraits.R
\name{print.traits.build}
\alias{print.traits.build}
\title{Generic for outputting a nice summary for austraits objects}
\usage{
\method{print}{traits.build}(x, ...)
}
\arguments{
\item{x}{traits.build database}

\item{\dots}{passed to print}
}
\value{
nicely printed table
}
\description{
Generic for outputting a nice summary for austraits objects
}
