% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_austraits.R
\name{print_austraits}
\alias{print_austraits}
\alias{print_austraits_traits}
\alias{print_austraits_taxa}
\title{Summarise counts for a particular variable of interest}
\usage{
print_austraits(austraits, var)

print_austraits_traits(austraits, var)

print_austraits_taxa(austraits, var)
}
\arguments{
\item{austraits}{A large list of tibbles built from austraits}

\item{var}{variable you use wish to see summary of}
}
\value{
dataframe of unique levels of variable with counts and percentage
}
\description{
Summarise counts for a particular variable of interest
}
\examples{
\dontrun{
print_austraits(austraits, "trait_name")
print_austraits(austraits, family)
}
}
