% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{print_versions}
\alias{print_versions}
\title{Print out AusTraits versions}
\usage{
print_versions(path, update = TRUE)
}
\arguments{
\item{path}{A file path, if AusTraits was previously downloaded direct to that folder}

\item{update}{Would you like the versions json be updated in case of new releases?}
}
\value{
A tibble containing version numbers and doi which can be used in load_austraits()
}
\description{
Print out AusTraits versions
}
