% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_unpack_combined_table.R
\name{recreate_contributors_dataframe}
\alias{recreate_contributors_dataframe}
\title{Recreate long-format contributors tibble}
\usage{
recreate_contributors_dataframe(combined_table_by_dataset)
}
\arguments{
\item{combined_table_by_dataset}{The combined table format of a traits.build built database.}
}
\value{
traits.build list object, but with additional fields (columns) appended to \code{traits} dataframe
}
\description{
The function \code{recreate_contributors_dataframe} takes the packed \code{data_collectors} column in the combined table output
of a traits.build database and reformats it into a long-format table, identical to what existed in the initial relational database.
}
\examples{
\dontrun{
}
}
