% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_unpack_combined_table.R
\name{recreate_methods_dataframe}
\alias{recreate_methods_dataframe}
\title{Recreate long-format methods tibble}
\usage{
recreate_methods_dataframe(combined_table_by_dataset)
}
\arguments{
\item{combined_table_by_dataset}{The combined table format of a traits.build built database.}
}
\value{
traits.build list object, but with additional fields (columns) appended to \code{traits} dataframe
}
\description{
The function \code{recreate_methods_dataframe} reformats the methods information included in
a traits.build combined output table into the original relational table format.
This mostly just requires selecting specific columns from the combined table,
but the data collectors column needs to be recreated.
}
\examples{
\dontrun{
}
}
