% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seperate_trait_values.R
\name{separate_trait_values}
\alias{separate_trait_values}
\title{Separate bounded trait values}
\usage{
separate_trait_values(data, definitions)
}
\arguments{
\item{data}{The trait data frame generated from austraits - see example}

\item{definitions}{The austraits definitions data frame}
}
\value{
trait tibble
}
\description{
This function reverts the action of bind_trait_values.
This function separates values that were concatenated so that studies that have multiple observations for a given trait will have seperate row for each observation.
}
\examples{
\dontrun{
traits <- austraits$traits \%>\% 
dplyr::filter(dataset_id == "Falster_2005_1")
traits
traits_bind <- bind_trait_values(traits)
separate_trait_values(traits_bind)
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
