% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_austraits.R
\name{summarise_database}
\alias{summarise_database}
\title{Summarise counts for a particular variable of interest}
\usage{
summarise_database(database, var)
}
\arguments{
\item{database}{traits.build database (list object)}

\item{var}{variable you use wish to see summary of (trait_name, genus, family)}
}
\value{
dataframe of unique levels of variable with counts and percentage
}
\description{
Summarise counts for a particular variable of interest
}
\examples{
\dontrun{
summarise_database(database = austraits, "trait_name")
summarise_database(database = austraits, "family")
}
}
