% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{trait_distribution_plot_numerical}
\alias{trait_distribution_plot_numerical}
\title{Title}
\usage{
trait_distribution_plot_numerical(
  austraits,
  plant_trait_name,
  y_axis_category,
  highlight = NA,
  hide_ids = FALSE
)
}
\arguments{
\item{austraits}{}

\item{plant_trait_name}{}

\item{y_axis_category}{}

\item{highlight}{}

\item{hide_ids}{}
}
\value{

}
\description{
Title
}
