% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_pivot_longer.R
\name{trait_pivot_longer}
\alias{trait_pivot_longer}
\title{Pivot wide format ausTrait data into a long format}
\usage{
trait_pivot_longer(data)
}
\arguments{
\item{data}{A tibble in wide format generated from trait_pivot_wider - see example}
}
\value{
A tibble in long format
}
\description{
trait_pivot_longer "gathers" wide format data into a "tidy" format
This function converts the data into long format where observations are on different rows and the type of observation is denoted by trait name
}
\examples{
\dontrun{
data <- austraits$traits \%>\% 
filter(dataset_id == "Falster_2003")
data #long format 
traits_wide <- trait_pivot_wider(data) 
traits_wide #wide format

values_long <- trait_pivot_longer(traits_wide, data$definitions)
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
