% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_unpack_combined_table.R
\name{unpack_context_properties}
\alias{unpack_context_properties}
\title{Unpack condensed context properties}
\usage{
unpack_context_properties(combined_table_by_dataset)
}
\arguments{
\item{combined_table_by_dataset}{The combined table format of a traits.build built database.}
}
\value{
traits.build list object, but with additional fields (columns) appended to \code{traits} dataframe
}
\description{
The function \code{unpack_context_properties} takes the 5 packed context properties columns in the combined table output
of a traits.build database and splits the information into separate columns, one for each context property.
}
\examples{
\dontrun{
}
}
