% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_unpack_combined_table.R
\name{unpack_data_contributors}
\alias{unpack_data_contributors}
\title{Unpack condensed data collectors}
\usage{
unpack_data_contributors(combined_table_by_dataset)
}
\arguments{
\item{combined_table_by_dataset}{The combined table format of a traits.build built database.}
}
\value{
traits.build list object, but with additional fields (columns) appended to \code{traits} dataframe
}
\description{
The function \code{unpack_data_collectors} takes the packed data collectors column in the combined table output
of a traits.build database and splits the information into separate columns, one for each data collector.
}
\examples{
\dontrun{
}
}
