% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_unpack_combined_table.R
\name{unpack_location_properties}
\alias{unpack_location_properties}
\title{Unpack condensed location properties}
\usage{
unpack_location_properties(combined_table)
}
\arguments{
\item{combined_table_by_dataset}{The combined table format of a traits.build built database.}
}
\value{
traits.build list object, but with additional fields (columns) appended to \code{traits} dataframe
}
\description{
The function \code{unpack_location_properties} takes the packed locations properties column in the combined table output
of a traits.build database and splits the information into separate columns, one for each location property.
}
\examples{
\dontrun{
}
}
