% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_df_to_list}
\alias{util_df_to_list}
\title{Convert dataframe to list}
\usage{
util_df_to_list(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A (yaml) list
}
\description{
Convert a dataframe to a named list, useful when converting to yaml.
}
\examples{
util_df_to_list(dplyr::starwars)
}
