% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_list_to_df2}
\alias{util_list_to_df2}
\title{Convert a list of lists to dataframe}
\usage{
util_list_to_df2(my_list, as_character = TRUE, on_empty = NA)
}
\arguments{
\item{my_list}{A list of lists to dataframe}

\item{as_character}{A logical value, indicating whether the values are read as character}

\item{on_empty}{Value to return if my_list is NULL, NA or is length == 0, default = NA}
}
\description{
Convert a list of lists to dataframe; requires that every list have same named elements.
}
\examples{
demo_list1 <- list(word1 = "this", word2 = "is", word3 = "an", word4 = "example", word5 = "list")
demo_list2 <- list(word1 = "and", word2 = "a", word3 = "second", word4 = "list", word5 = "also")
combined_list <- list(demo_list1, demo_list2)
util_list_to_df2(combined_list)
}
