% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_wide_table.R
\name{as_wide_table}
\alias{as_wide_table}
\title{Create a single wide table from a traits.build data object}
\usage{
as_wide_table(database)
}
\arguments{
\item{database}{traits.build database (list object)}
}
\value{
A single wide table with collapsed contexts and locations text and with
some cols renamed for alignment with other resources
}
\description{
Create a single wide table from a traits.build data object
}
\examples{
\dontrun{
austraits \%>\% as_wide_table()
}
}
