% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austraits-package.R
\docType{package}
\name{austraits-package}
\alias{austraits}
\alias{austraits-package}
\title{austraits: Helpful functions to access, summarise and wrangle austraits data}
\description{
austraits is a package for accessing, summarising and wrangling austraits data. The package allows users to explore the scope and type of data available, as well as subset and wrangle relevant specific data for furture research. This package is still under development. See README and vignette('austraits') for further information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://traitecoevo.github.io/austraits/}
  \item Report bugs at \url{https://github.com/traitecoevo/austraits/issues}
}

}
\author{
\strong{Maintainer}: Fonti Kar \email{f.kar@unsw.edu.au} (\href{https://orcid.org/0000-0002-2760-3974}{ORCID})

Authors:
\itemize{
  \item Daniel Falster \email{daniel.falster@unsw.edu.au} (\href{https://orcid.org/0000-0002-9814-092X}{ORCID})
  \item Dony Indiarto (\href{https://orcid.org/0000-0001-9546-8201}{ORCID})
}

}
\keyword{internal}
