% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_databases.R
\name{bind_databases}
\alias{bind_databases}
\title{Bind multiple traits.build data objects into a single data object}
\usage{
bind_databases(..., d = list(...))
}
\arguments{
\item{...}{Arguments passed to other functions}

\item{List}{of traits.build databases to bind together (lists of tibbles)}
}
\value{
Compiled database as a single large list
}
\description{
\code{bind_databases} binds all the listed studies into a single traits.build
database object as a large list.
}
