% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{check_compatibility}
\alias{check_compatibility}
\title{Check compatibility}
\usage{
check_compatibility(austraits)
}
\arguments{
\item{austraits}{dataframe generated by traits build}
}
\value{
logical (T/F) output and messaging for uncompatible versions
}
\description{
Function to check whether the data object has been compiled by the traits.build workflow and
therefore has a data structure that is appropriate for use with austraits functions.
}
\examples{
\dontrun{
check_compatibility(austraits)
}
}
\author{
Elizabeth Wenk - e.wenk@unsw.edu.au
}
