% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{check_traits_compatibility}
\alias{check_traits_compatibility}
\title{Check compatibility of traits table}
\usage{
check_traits_compatibility(trait_data)
}
\arguments{
\item{trait_data}{the traits table in a traits.build database}
}
\value{
logical, TRUE indicating version traits table came from traits.build version > 1.0
}
\description{
Check compatibility of traits table
}
