% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dataset.R
\name{extract_dataset}
\alias{extract_dataset}
\title{Extract all data for a particular dataset}
\usage{
extract_dataset(database, dataset_id)
}
\arguments{
\item{database}{traits.build database (list object)}

\item{dataset_id}{character string that matches a \code{dataset_id} in the database}
}
\value{
List of tibbles containing all traits.build data and metadata for the specified dataset(s).
}
\description{
Function to subset all data associated with a particular dataset from a traits.build relational database.
}
\details{
\code{extract_dataset} has been developed to extract data for specific datasets from databases built using the traits.build workflow.
Learn more at:
\url{https://github.com/traitecoevo/traits.build} &
\url{https://github.com/traitecoevo/traits.build-book}

Note to AusTraits users:
\itemize{
\item This function works with AusTraits version >= 5.0.0 (from Nov 2023 release)
\item For AusTraits versions <= 4.2.0 (up to Sept 2023 release) see \url{https://github.com/traitecoevo/austraits} for how to install old versions of the package or download a newer version of the database.
}
}
\examples{
\dontrun{
extract_dataset(database, "Falster_2003")
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
