% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dataset.R
\name{extract_dataset}
\alias{extract_dataset}
\title{Extract all data for a particular dataset}
\usage{
extract_dataset(austraits, dataset_id)
}
\arguments{
\item{austraits}{\itemize{
\item A large list of tibbles built from austraits
}}

\item{dataset_id}{\itemize{
\item character string that matches a dataset_id in the data
}}
}
\value{
A large list of tibbles containing all austraits information for one particular dataset
}
\description{
Function to subset of all data associated with a particular dataset from austraits
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
