% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxa.R
\name{extract_taxa}
\alias{extract_taxa}
\title{Extract data for one specific taxa}
\usage{
extract_taxa(austraits, family, genus)
}
\arguments{
\item{austraits}{\itemize{
\item A large list of tibbles built from austraits
}}

\item{family}{\itemize{
\item character string of family
}}

\item{genus}{\itemize{
\item character string of genus
}}

\item{taxon_name}{\itemize{
\item character string of taxon name
}}
}
\value{
A large list of tibbles containing all austraits information for specificied taxa
}
\description{
Function to subset of all data associated with a particular dataset from austraits
}
\examples{
\dontrun{
extract_taxa(austraits, family = "Proteaceae")
extract_taxa(austraits, genus = "Acacia")
}
}
\author{
Fonti Kar - f.kar@unsw.edu.au
}
