% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxa.R
\name{extract_taxa}
\alias{extract_taxa}
\title{Extract all data for specific taxa}
\usage{
extract_taxa(database, family = NULL, genus = NULL, taxon_name = NULL)
}
\arguments{
\item{database}{traits.build database (list object)}

\item{family}{character string of family or families}

\item{genus}{character string of genus or genera}

\item{taxon_name}{character string of taxon name(s)}
}
\value{
List of tibbles containing all traits.build data and metadata for the specified taxa.
}
\description{
Function to subset of all data associated with a particular taxon from a traits.build relational database.
}
\details{
\code{extract_taxa} has been developed to extract data for specific taxa from databases built using the traits.build workflow.
Learn more at:
\href{https://github.com/traitecoevo/traits.build-book}{https://github.com/traitecoevo/traits.build} &
\url{https://github.com/traitecoevo/traits.build-book}

Note to AusTraits users:
\itemize{
\item This function works with AusTraits version >= 5.0.0 (from Nov 2023 release)
\item For AusTraits versions <= 4.2.0 (up to Sept 2023 release) see \url{https://github.com/traitecoevo/database} for how to install old versions of the package or download a newer version of the database.
}
}
\examples{
\dontrun{
extract_taxa(database = austraits, family = "Proteaceae")
extract_taxa(database = austraits, genus = "Acacia")
}
}
\author{
Fonti Kar - f.kar@unsw.edu.au
}
