% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trait.R
\name{extract_trait}
\alias{extract_trait}
\title{Extract data for specific traits}
\usage{
extract_trait(austraits, trait_names, taxon_names = NULL),
}
\arguments{
\item{austraits}{\itemize{
\item A large list of tibbles built from austraits
}}

\item{trait_names}{\itemize{
\item character string of trait that will be extracted
}}

\item{taxon_names}{\itemize{
\item optional arugument
}}
}
\value{
A large list of tibbles containing all austraits information for one particular dataset
}
\description{
Function to subset of all data associated with a particular dataset from austraits
}
\examples{
\dontrun{
extract_dataset(austraits, "Falster_2003")
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
