% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_traits_table}
\alias{get_traits_table}
\title{Retrieve traits table if user passes traits.build object.}
\usage{
get_traits_table(database)
}
\arguments{
\item{database}{traits.build database or traits table in a traits.build database}
}
\description{
Retrieve traits table if user passes traits.build object.
}
