% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_all}
\alias{join_all}
\alias{join_taxonomy}
\alias{join_methods}
\alias{join_locations}
\alias{join_sites}
\alias{join_contexts}
\title{Join study details into main \code{traits} dataset}
\usage{
join_all(austraits)

join_taxonomy(austraits, ...)

join_methods(
  austraits,
  vars = c("methods", "year_collected_start", "year_collected_end", "collection_type")
)

join_locations(austraits, ...)

join_contexts(
  austraits,
  vars = c("dataset_id", "context_name", "context_property", "value")
)
}
\arguments{
\item{austraits}{dataframe generated by austraits build}

\item{...}{arguments passed to vars to select relevant column names from their respective tables}
}
\value{
austraits list object, but with additional variables appended to \code{traits} dataframe
}
\description{
Function to append all study information (method, location, taxonomic, context) variables into trait database
}
\examples{
\dontrun{
austraits$traits

#Append sites data
(austraits \%>\% join_sites)$traits

#Append contexts
(austraits \%>\% join_contexts)$traits

# Append methods
(austraits \%>\% join_methods)$traits

#Append taxonomic details
(austraits \%>\% join_taxonomy)$traits

#Append all information
(austraits \%>\% join_all)$traits
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
