% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_location_coordinates}
\alias{join_location_coordinates}
\title{Joining location coordinates to traits table}
\usage{
join_location_coordinates(austraits)
}
\arguments{
\item{austraits}{traits.build database}
}
\value{
traits.build list object, but with additional fields (columns)
for latitude and longitude appended to \code{traits} dataframe
}
\description{
Function to merge geographic coordinates (latitude/longitude)
stored in the locations table of a traits.build database into the core traits table.
}
\details{
the \code{join_} functions have been developed to join relational tables for databases built using the traits.build workflow.
Learn more at:
\url{https://github.com/traitecoevo/traits.build} &
\url{https://github.com/traitecoevo/traits.build-book}

Note to AusTraits users:
\itemize{
\item This function works with AusTraits version >= 5.0.0 (from Nov 2023 release)
\item For AusTraits versions <= 4.2.0 (up to Sept 2023 release) see \url{https://github.com/traitecoevo/austraits}  for how to install old versions of the package or download a newer version of the database.
}
}
\examples{
\dontrun{
(austraits \%>\% join_location_coordinates)$traits
}

}
