% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_sites}
\alias{join_sites}
\alias{join_locations}
\title{Joining location info for AusTraits versions <= 3.0.2 (Deprecated)}
\usage{
join_sites(austraits, vars = c("longitude (deg)", "latitude (deg)"))
}
\description{
Joining location info for AusTraits versions <= 3.0.2 (Deprecated)
}
