% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all.R
\name{join_sites}
\alias{join_sites}
\title{Joining location info for AusTraits versions <= 3.0.2}
\usage{
join_sites(austraits, vars = c("longitude (deg)", "latitude (deg)"))
}
\arguments{
\item{austraits}{dataframe generated by austraits build}

\item{vars}{variables to select from the respective table where information is being joined from. Not available for contexts table}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Joining location info for AusTraits versions <= 3.0.2
}
