% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{load_austraits}
\alias{load_austraits}
\title{Load AusTraits database into R console}
\usage{
load_austraits(
  doi = NULL,
  version = NULL,
  path = "data/austraits",
  update = FALSE
)
}
\arguments{
\item{doi}{character string - doi of particular version}

\item{version}{character string - version number of database}

\item{path}{file path to where AusTraits will be downloaded}

\item{update}{if TRUE, AusTraits versions .json will be re-downloaded}
}
\value{
a large list containing AusTraits data tables
}
\description{
Load AusTraits database into R console
}
\examples{
\dontrun{
austraits <- load_austraits(version = "3.0.2", path = "data/austraits")
}
}
\seealso{
get_versions get_version_latest
}
