% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{load_austraits}
\alias{load_austraits}
\title{Load AusTraits}
\usage{
load_austraits(path = NA)
}
\arguments{
\item{path}{file path to where downloaded AusTraits .rds file is found}
}
\value{
AusTraits data
}
\description{
Read in AusTraits .rds file into R, if filepath not specified, it reads in the lite version used for testing
}
\examples{
\dontrun{
austraits <- load_austraits() #loads Lite version for testing
austraits <- load_austraits("data/austraits-2.1.0.rds")
}

}
