% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{load_austraits}
\alias{load_austraits}
\title{Load AusTraits}
\usage{
load_austraits(path)
}
\arguments{
\item{path}{file path to where downloaded AusTraits .rds file is found, if empty function will create an 'austraits' directory in your data folder}
}
\value{
AusTraits data
}
\description{
Downloads AusTraits version 2.1.0 and reads into R
}
\examples{
\dontrun{
austraits_lite <- load_austraits_lite() #loads Lite version for testing
austraits <- load_austraits("data/austraits/austraits-2.1.0.rds")
}
}
