% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_austraits.R
\name{load_austraits}
\alias{load_austraits}
\title{Load AusTraits}
\usage{
load_austraits(path)
}
\arguments{
\item{path}{file path to where download data. By default set to "data/austraits"}

\item{link}{link to Zenodo record to download}

\item{update}{should files be downloaded anew?}
}
\value{
AusTraits data object
}
\description{
Downloads AusTraits version 3.0.02 and reads into R
}
\examples{
\dontrun{
austraits_lite <- load_austraits_lite() #loads Lite version for testing
austraits <- load_austraits()
}
}
