% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_.R
\name{lookup_location_property}
\alias{lookup_location_property}
\title{Look up location properties}
\usage{
lookup_location_property(database, term)
}
\arguments{
\item{database}{traits.build database (list object)}

\item{term}{character string for location property search term}
}
\value{
vector containing location properties that contains search term
}
\description{
Look up location properties that contain a specific search term.
}
\examples{
\dontrun{
austraits \%>\% lookup_location_property("soil")
}
}
