% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_.R
\name{lookup_trait}
\alias{lookup_trait}
\title{Look up a particular trait term}
\usage{
lookup_trait(database, term)
}
\arguments{
\item{database}{traits.build database (list object)}

\item{term}{character string for trait search term}
}
\value{
vector containing traits that contains search term
}
\description{
Look up a particular trait term
}
\examples{
\dontrun{
austraits \%>\% lookup_trait("leaf") \%>\% extract_trait(database = austraits, .)
}
}
