% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locations.R
\name{plot_site_locations}
\alias{plot_site_locations}
\title{Produce location maps of trait values (Deprecated)}
\usage{
plot_site_locations(austraits, feature = "trait_name", ...)
}
\description{
Produce location maps of trait values (Deprecated)
}
