% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_site_locations.R
\name{plot_site_locations}
\alias{plot_site_locations}
\title{Produce site maps}
\usage{
plot_site_locations(
  traits,
  feature = "trait_name",
  size = 0.5,
  alpha = 0.8,
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{traits}{trait dataframe generated from austraits with site details appended. See join_all and examples}

\item{feature}{grouping/classification categories e.g trait_name, collection_type}
}
\value{
ggplot of sites
}
\description{
Plot location where trait data was collected from
}
\examples{
\dontrun{
#All traits from a given study
data <- austraits \%>\% extract_dataset(dataset_id = "Falster_2003") \%>\% join_all() 
data$traits \%>\% plot_site_locations("trait_name")

#Single trait
data <- austraits \%>\% extract_trait(trait_names = c("plant_height")) \%>\% join_all() 
data$traits \%>\% plot_site_locations("trait_name")
}
}
\author{
Dony Indiarto - d.indiarto@student.unsw.edu.au
}
