% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locations.R
\name{plot_site_locations}
\alias{plot_site_locations}
\title{Produce location maps of trait values}
\usage{
plot_site_locations(traits, feature = "trait_name", ...)
}
\arguments{
\item{traits}{traits table with site details appended. See join_location_coordinates and examples}

\item{feature}{grouping/classification categories e.g trait_name, collection_type for <= v3.0.2}

\item{...}{arguments passed to ggplot()}
}
\value{
ggplot of sites
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Plot location where trait data was collected from
}
\author{
Dony Indiarto - d.indiarto@student.unsw.edu.au
}
