% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_site_locations.R
\name{plot_site_locations}
\alias{plot_site_locations}
\title{Produce site maps}
\usage{
plot_site_locations(
  traits,
  feature = "trait_name",
  size = 0.5,
  alpha = 0.8,
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{feature}{grouping/classification categories}

\item{trait}{trait dataframe generated from austraits with site details appended. See join_all and examples}
}
\value{
ggplot of sites
}
\description{
Plot
}
\examples{
\dontrun{
data <- austraits \%>\% join_all() \%>\% `[[`("traits") 
data \%>\% plot_site_locations()

data <- austraits \%>\% extract_trait(trait_names = c("leaf_angle", "leaf_area", "specific_leaf_area", "wood_density", "plant_height")) \%>\% join_all() \%>\% `[[`("traits")
data \%>\% plot_site_locations("trait_name")
}
}
\author{
Dony Indiarto - d.indiarto@student.unsw.edu.au
}
