% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locations.R
\name{plot_site_locations}
\alias{plot_site_locations}
\title{Produce location maps of trait values (Deprecated)}
\usage{
plot_site_locations(traits, feature = "trait_name", ...)
}
\arguments{
\item{traits}{traits table with site details appended. See join_all and examples}

\item{feature}{grouping/classification categories e.g trait_name, collection_type for <= v3.0.2}

\item{...}{arguments passed to ggplot()}
}
\value{
ggplot of sites
}
\description{
Plot location where trait data was collected from
}
\author{
Dony Indiarto - d.indiarto@student.unsw.edu.au
}
