% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.austraits.R
\name{print.austraits}
\alias{print.austraits}
\title{Generic for outputting a nice summary for austraits objects}
\usage{
\method{print}{austraits}(x, ...)
}
\arguments{
\item{x}{austraits list object}

\item{\dots}{passed to print}
}
\value{
nicely printed table
}
\description{
Generic for outputting a nice summary for austraits objects
}
