% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_trait_values.R
\name{separate_trait_values}
\alias{separate_trait_values}
\title{Separate bounded trait values}
\usage{
separate_trait_values(trait_data, definitions)
}
\arguments{
\item{trait_data}{The traits table in a traits.build database - see example}

\item{definitions}{The austraits definitions data frame}
}
\value{
trait tibble
}
\description{
This function reverts the action of bind_trait_values.
This function separates values that were concatenated so that studies that have multiple observations for a given trait will have separate row for each observation.
}
\examples{
\dontrun{
trait_data <- austraits$traits \%>\% 
dplyr::filter(dataset_id == "Falster_2005_1")
trait_data
traits_bind <- bind_trait_values(trait_data)
separate_trait_values(traits_bind)
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
