% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_austraits.R
\name{summarise_austraits}
\alias{summarise_austraits}
\title{Summarise counts for a particular variable of interest}
\usage{
summarise_austraits(austraits, var)
}
\arguments{
\item{austraits}{A large list of tibbles built from austraits}

\item{var}{variable you use wish to see summary of (trait_name, genus, family)}
}
\value{
dataframe of unique levels of variable with counts and percentage
}
\description{
Summarise counts for a particular variable of interest
}
\examples{
\dontrun{
summarise_austraits(austraits, "trait_name")
summarise_austraits(austraits, "family")
}
}
