% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_pivot_longer.R
\name{trait_pivot_longer}
\alias{trait_pivot_longer}
\title{Pivot wide format traits table into long format}
\usage{
trait_pivot_longer(wide_data)
}
\arguments{
\item{wide_data}{output from trait_pivot_wider.}
}
\value{
A tibble in long format
}
\description{
Function to "gather" wide format data into a "tidy" format.

This function reverts the actions of the function austraits::trait_pivot_wider.

It begins with a derivation of a traits.build traits table, where multiple measurements that comprise a single observation are displayed on a single row,with a column for each trait. It then converts the table into long format where measurements of multiple traits that comprise a single observation are on different rows and a column specifying the trait names is added.
}
\details{
\code{trait_pivot_longer} has been developed to pivot the traits table for a database build using the traits.build workflow.
Learn more at:
\link{https://github.com/traitecoevo/traits.build} &
\link{https://github.com/traitecoevo/traits.build-book}

Note to AusTraits users:
\itemize{
\item This function works with AusTraits version >= 5.0.0 (from Nov 2023 release)
\item For AusTraits versions <= 4.2.0 (up to Sept 2023 release) see \link{https://github.com/traitecoevo/austraits} for how to install old versions of the package or download a newer version of the database."
}
}
\examples{
\dontrun{
data <- austraits$traits \%>\% 
filter(dataset_id == "Falster_2003")
data #long format 
traits_wide <- trait_pivot_wider(data) 
traits_wide #wide format

values_long <- trait_pivot_longer(traits_wide)
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au

Fonti Kar - fonti.kar@unsw.edu.au
}
