% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_pivot_wider.R
\name{trait_pivot_wider}
\alias{trait_pivot_wider}
\title{Pivot long format ausTrait data into a wide format}
\usage{
trait_pivot_wider(data)
}
\arguments{
\item{data}{A data object g}
}
\value{
data
}
\description{
trait_pivot_wider "widens" long format data ("tidy data")
Long format data has measurements on different rows, and just a single column which denotes what type of measurement it is.
This function manipulates the data frame into wide format has data so that each trait in it's own column.
}
\examples{
data <- austraits$traits \%>\% 
filter(dataset_id == "Falster_2003")
data #long format 
traits_wide <- trait_pivot_wider(data) 
traits_wide #wide format
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
