% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_pivot_wider.R
\name{trait_pivot_wider}
\alias{trait_pivot_wider}
\title{Pivot long format ausTrait data into a wide format}
\usage{
trait_pivot_wider(data)
}
\arguments{
\item{data}{A tibble generated from ausTraits - see example}
}
\value{
list of tibbles in wide format
}
\description{
trait_pivot_wider "widens" long format data ("tidy data")
Long format data has measurements on different rows and the type of measurement denoted in a single column.
This function manipulates the data into wide format so that each trait in it's own column.
}
\examples{
\dontrun{
data <- austraits$traits \%>\% 
filter(dataset_id == "Falster_2003")
data #long format 
traits_wide <- trait_pivot_wider(data) 
traits_wide #wide format
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
